fix: Resolve todo UI update issues and test failures
- Fix desktop platform Firebase login fallback logic
- Resolve web notification platform compatibility issues
- Update AI generator widgets to use full width layout
- Center align AI generator header text
- Fix test timeouts and assertion failures
- Update MockTodoRepository with missing interface methods
- Improve Firebase sync service type safety
